준석_ 페이징처리, 로그인한 유저에 따라서 내가 작성한 리뷰 다르게 출력하기 완료.

+++ Mealchelin/src/main/java/com/mealchelin/mvc/review/model/vo/Review.java
@@ -26,7 +26,7 @@ public class Review {
 	    
 	    private String status; // 상태 (Y/N)
 	    
-	    private Long userNo; // 회원 번호
+	    private int userNo; // 회원 번호
 	    
 	    private Long prdNo; // 상품 번호
